+++ gems/aws-sdk-core/lib/aws-sdk-core/json/error_handler.rb
@@ -2,10 +2,8 @@
 
 module Aws
   module Json
-    class ErrorHandler < Xml::ErrorHandler
+    class ErrorHandler < Aws::ErrorHandler
 
-      # @param [Seahorse::Client::RequestContext] context
-      # @return [Seahorse::Client::Response]
       def call(context)
         @handler.call(context).on(300..599) do |response|
           response.error = error(context)
@@ -19,7 +17,7 @@ module Aws
         json = Json.load(body)
         code = error_code(json, context)
         message = error_message(code, json)
-        data = parse_error_data(context, code)
+        data = parse_error_data(context, body, code)
         [code, message, data]
       rescue Json::ParseError
         [http_status_error_code(context), '', EmptyStructure.new]
@@ -48,7 +46,7 @@ module Aws
       end
 
       def remove_prefix(error_code, context)
-        if prefix = context.config.api.metadata['errorPrefix']
+        if (prefix = context.config.api.metadata['errorPrefix'])
           error_code.sub(/^#{prefix}/, '')
         else
           error_code
@@ -63,9 +61,9 @@ module Aws
         end
       end
 
-      def parse_error_data(context, code)
+      def parse_error_data(context, body, code)
         data = EmptyStructure.new
-        if error_rules = context.operation.errors
+        if (error_rules = context.operation.errors)
           error_rules.each do |rule|
             # match modeled shape name with the type(code) only
             # some type(code) might contains invalid characters
@@ -73,7 +71,7 @@ module Aws
             match = rule.shape.name == code.gsub(/[^^a-zA-Z0-9]/, '')
             next unless match && rule.shape.members.any?
 
-            data = Parser.new(rule).parse(context.http_response.body_contents)
+            data = Parser.new(rule).parse(body)
             # errors support HTTP bindings
             apply_error_headers(rule, context, data)
           end