+++ gems/aws-sdk-core/lib/aws-sdk-core/plugins/protocols/json_rpc.rb
@@ -5,18 +5,17 @@ module Aws
     module Protocols
       class JsonRpc < Seahorse::Client::Plugin
 
+        option(:protocol, 'json')
+
         option(:simple_json,
           default: false,
           doc_type: 'Boolean',
           docstring: <<-DOCS)
 Disables request parameter conversion, validation, and formatting.
-Also disable response data type conversions. This option is useful
-when you want to ensure the highest level of performance by
-avoiding overhead of walking request parameters and response data
-structures.
-
-When `:simple_json` is enabled, the request parameters hash must
-be formatted exactly as the DynamoDB API expects.
+Also disables response data type conversions. The request parameters
+hash must be formatted exactly as the API expects.This option is useful
+when you want to ensure the highest level of performance by avoiding
+overhead of walking request parameters and response data structures.
           DOCS
 
         option(:validate_params) { |config| !config.simple_json }
@@ -24,7 +23,6 @@ be formatted exactly as the DynamoDB API expects.
         option(:convert_params) { |config| !config.simple_json }
 
         handler(Json::Handler)
-
         handler(Json::ErrorHandler, step: :sign)
 
       end