+++ gems/aws-sdk-core/lib/aws-sdk-core/plugins/protocols/query.rb
@@ -1,13 +1,15 @@
 # frozen_string_literal: true
 
-require_relative '../../query'
-
 module Aws
   module Plugins
     module Protocols
       class Query < Seahorse::Client::Plugin
+
+        option(:protocol, 'query')
+
         handler(Aws::Query::Handler)
         handler(Xml::ErrorHandler, step: :sign)
+
       end
     end
   end