+++ gems/aws-sdk-core/lib/aws-sdk-core/plugins/protocols/rest_json.rb
@@ -4,13 +4,14 @@ module Aws
   module Plugins
     module Protocols
       class RestJson < Seahorse::Client::Plugin
+
+        option(:protocol, 'rest-json')
+
         handler(Rest::Handler)
-        # Rest::Handler will set a default JSON body, so size can be checked
-        # if this handler is run after serialization.
         handler(Rest::ContentTypeHandler, priority: 30)
         handler(Json::ErrorHandler, step: :sign)
-      end
 
+      end
     end
   end
 end