+++ gems/aws-sdk-core/CHANGELOG.md
@@ -1,6 +1,12 @@
 Unreleased Changes
 ------------------
 
+* Feature - Support RpcV2 protocol.
+
+* Feature - Add CBOR encoder and decoder.
+
+* Issue - Enhance, refactor, and rebase protocols to be consistent. Ensure protocol tests are run for each engine.
+
 3.198.0 (2024-06-24)
 ------------------